Add files via upload
In this repository, I have performed an in-depth outlier detection and handling analysis on two distinct datasets, placement.csv and weight-height.csv.

For the 

placement.csv dataset, I used both the Z-score and IQR (Interquartile Range) methods to identify and address outliers in the cgpa column. The analysis involved two primary approaches: 

trimming and capping. Trimming involved removing the identified outlier rows, which in this case reduced the dataset from 1000 rows to 995. The capping method, also known as winsorization, was applied to replace the outlier values with the upper and lower boundary limits, calculated as 

mean +/- 3 * std, ensuring the dataset maintained its original size.

Similarly, for the 

weight-height.csv dataset, I applied the percentile outlier detection method to the Height column. The upper and lower limits for outliers were determined by the 99th and 1st percentiles, respectively. As with the previous dataset, both trimming and capping methods were used. Trimming removed the identified outliers, resulting in a reduction from 10,000 rows to 9,799 rows. The capping method, on the other hand, replaced the extreme values with the percentile limits, preserving the dataset's size. This comprehensive approach to outlier detection and handling provides a clear demonstration of different techniques and their impact on data integrity and distribution.